[Bronze V] Title: 꼬마 정민, Time: 0 ms, Memory: 1124 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main(void){
-    long long A, B, C;
+    long A, B, C;
     
-    scanf("%lld %lld %lld", &A, &B, &C);
-    printf("%lld", A+B+C);
+    scanf("%ld %ld %ld", &A, &B, &C);
+    printf("%ld", A+B+C);
    
     return 0;
 }
